refactor: Associate no modder with classes
